test-webhook/requirements.txt
functions-framework==3.*
google-auth
google-auth-httplib2
requests
flask